backend/testing/a2a_learning/pyproject.toml
[project]
name = "a2a-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "strands-agents",
  "strands-agents[a2a]"
]
